+++ Rebus.RabbitMq/RabbitMqTransport.cs
@@ -185,7 +185,7 @@ namespace Rebus.RabbitMq
             }
         }
 
-        public async Task<TransportMessage> Receive(ITransactionContext context, CancellationToken cToken = default(CancellationToken))
+        public async Task<TransportMessage> Receive(ITransactionContext context, CancellationToken cancellationToken = default(CancellationToken))
         {
             if (Address == null)
             {