+++ Rebus/Transport/Msmq/MsmqTransport.cs
@@ -5,6 +5,7 @@ using System.IO;
 using System.Messaging;
 using System.Runtime.Serialization;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 using Rebus.Bus;
@@ -160,7 +161,7 @@ namespace Rebus.Transport.Msmq
         /// it under the <see cref="CurrentTransactionKey"/> key in the given <paramref name="context"/>. If one already exists, an exception will be thrown
         /// (because we should never have to receive multiple messages in the same transaction)
         /// </summary>
-        public async Task<TransportMessage> Receive(ITransactionContext context, CancellationToken cToken = default(CancellationToken))
+        public async Task<TransportMessage> Receive(ITransactionContext context, CancellationToken cancellationToken = default(CancellationToken))
         {
             if (context == null) throw new ArgumentNullException(nameof(context));
             if (_inputQueueName == null)
@@ -198,7 +199,7 @@ namespace Rebus.Transport.Msmq
                 var headers = _extensionSerializer.Deserialize(message.Extension, message.Id);
                 var body = new byte[message.BodyStream.Length];
 
-                await message.BodyStream.ReadAsync(body, 0, body.Length);
+                await message.BodyStream.ReadAsync(body, 0, body.Length, cancellationToken);
 
                 return new TransportMessage(headers, body);
             }